#13 Updating Tokenizer to provide a nicer interface for mapped values

+++ src/Bank/Money/Money.py
@@ -1,16 +1,16 @@
 class Money:
     """Money object"""
 
-    def __init__(self, amount: int, currency: str):
+    def __init__(self, amount: float, currency: str):
         """ Money."""
         self.__amount = amount
         self.__currency = currency.upper()
 
-    def add(self, amount: int) -> None:
+    def add(self, amount: float) -> None:
         """ Adds amount to the money """
         self.__amount = amount + amount
 
-    def multiply(self, multiplier: int) -> None:
+    def multiply(self, multiplier: float) -> None:
         """ Multiplies money by an amount"""
         self.__amount = self.__amount * multiplier
 
@@ -18,7 +18,7 @@ class Money:
         """ Returns the currency """
         return self.__currency
 
-    def get_amount(self) -> int:
+    def get_amount(self) -> float:
         """Returns the amount"""
         return self.__amount
